SLM-231: Added wiremock stubs for One Time Code API endpoints
